binboard: rewrite to use pieces and iv
Not too much new functionality in terms of `iv` - but it makes accessing
specific indices way easier. The real upgrade is using an array of Ints
instead of just an Int - now, we aren't just having each position on the
board be a single bit, but instead an actual piece. In the future, I
might skip the middleman and just make it an array of pieces. Who needs
the integer form? Anyways, we're building up the foundations for an
actual chess board.