coord/from_algebraic: init
We'll want to be able to handle moves with algebraic notation. Just a
simple parsing function and a basic test. The code is quite pretty,
though! I think this is the best I've handled errors yet. The big
improvement I want to make is to start defining custom error types, to
make testing easier since I don't have to match on strings.