coord/from_pair: rewrite to use `result`
Safer, since now if we send in an invalid row/col, we'll get a clear
error to understand what's going on. We also add a test to ensure it's
working correctly and prevent issues in the future.

We also add `to_algebraic` (I would've split it out into its own
commit, but it was in the middle of everything). It's not checked for
input errors, under the assumption that the user will create their
coordinates via `from_pair`, which *is* checked for errors. I probably
should just check for errors anyways, but I'll wait for it to hurt me,
and then complain.