coord: rename to position and revamp
I'm planning on writing a wrapper clsas around position. Before that,
though, I want to rewrite position to be more sane. Now, it simply
handles a row and a column - the index can be accessed via a function.
We also get rid of the index-based creation, meaning we can now rename
`from_pair` to just `new`. The coming `square` class will wrap around
position and piece, and I then plan on revamping the board to instead be
an array of positions. Who needs performance?