board: add `set` and don't store rows/cols
For my chess implementation, we can expect a board to always be of
length and width 8. However, not being able to make custom width/lengths
makes testing slightly more inconvenient, since the board is expected to
be 8x8 now. We add the functions `empty` and `create` to make QOL here
easier. `create` takes an input dataset, validates that it's the right
length, and returns it. `empty` simply returns an empty board, which you
can then modify.

Speaking of modifying the data, we add a `set` function, similar to the
`get` function. To minimize code repetition, we leverage a helper
function `pos_is_valid` to ensure the position will work. This helper
function should be moved to the `position` module in the future, but
before that, we need a way to standardize width and length 8 between
modules, so that's for another day.