board/from_fen: init
Big advances here! We can now parse the primary part of a fen string:
the one containing the board state. This seems to work for all the
strings I've tried. I'm technically missing some validation on the fen
string being too long or too short, but the core logic is implemented,
mainly from scratch!

We utilize our new helper function, `choose.cases`, to lower
indentation. The function is literally just a wrapper around `case`, but
because of `use`, it lets me decrease indentation. Love it!