tests: rewrite to use birdie and use `let assert`
Birdie is great bc it encodes the expected outputs for me - I don't need
to paste in what the values *should* be. In a similar vein, `let assert`
is actually way better for dealing with test failures - `try` wasn't
failing the test if the result was an Error, since it never got to the
actual test. I was planning on writing a function to handle this myself,
but `let assert` seems to do what I want. Also heavily leveraging
`string.inspect`, since Birdie needs strings.