flake: split into `flake.nix` and `shell.nix`
The shell.nix won't really work very well on its own - I'm not good
enough at the non-flake way of things yet. But it does make the
flake.nix file lighter - it can now simply be dedicated to grabbing the
custom packages from the `nixPackages/` directory, and leave the
devshell stuff to the shell.nix.