position: make index in relation to player
Found a bug that I hadn't noticed for awhile - the index returned from
the top left, but all my functions relying on the board's index expected
it to start in the bottom left, to be in reference to the current
player. To deal with this, we replace `get_index` with `to_player_index`
- I don't think it makes sense to still name it something that's vague
about what *kind* of index it is. We also make the `get_pos` test
actually test a board that isn't symmetrical, and hopefully notice these
kinds of issues in the future. Our tests should always be on a realistic
board, not just an empty one.