board: add `get_pieces` function
If I'm going to be doing move logic, I need a way to get all the pieces
of a color, and their positions. At first, I started rewriting the board
module to be a list of Squares, each one containing a piece and its
position. But, that's silly, and would require rewriting basically
everything. However, the Square idea was good - it should just be used
at runtime. We add some helper functions to turn a board index into a
position, and use index_map to map each piece on the board to a square,
containing that piece and its position. Finally, of course, some tests.
This seems great, and will let us (finally) start implementing some real
move logic.

The moment I said that, I remembered that I wanted to add Board
functions for getting a row and column, now that we have Rank and File
split into their own types... and I should probably also make one for
diagonals... and I don't know how I should do that. Hopefully someone
else has thought of this and I can reference their implementation. Much
to do. But, even though progress feels slow, and much of this will
likely be undone when I actually get to writing an implementation that's
 *fast*, we are making steady headway.