board: add `obstructed_distance` function
Returns the distance in a direction until you either hit a wall or hit
another piece. If you hit an enemy piece, it includes that as a valid
space, since you could capture it.

We also get rid of the `distance_to_edge()` function -
`obstructed_distance` ends up doing its functionality and *more*, so the
original doesn't really make sense anymore.

Finally, we add some basic tests to ensure that the function is working
somewhat sensibly! It's not the 8 tests we had before, but it's a basic
ensuring of functionality.