square, board/get_pieces: remove
I don't think this should be part of the board class. We can re-add it
later if I change my mind - but I don't think it's very useful with how
I see this developing at the current moment. Its functionality can be
part of a larger function for iterating through pieces and getting
moves.