direction: rename to SlidingDirection
Spent a lot of time thinking about whether I wanted to rework the code
to use offsets rather than vague directions. But, after thinking about
it, I think this style works better for the sliding pieces - it doesn't
require a recursive function, and can instead work purely off indices.
However, it's really not built for something like a knight - and a pawn
requires more work from something like a Game type, to check its en
passant status, and whether we're on the first move. So, we rename the
Direction type to SlidingDirection, to clarify its purpose. Ideally, I
would have a SlidingPiece type, with only some pieces, for handling
errors up top. But that's a commit for another day.