move: add more constructors
Fine, "Records". Whatever the functional programming name is for them.
Regardless, this is super helpful for representing different *types* of
moves -- and we can add more anytime. This also makes the UX much better
for applying moves to a board, and then having the game use that board -
since now `apply` takes in a game, you never have to work with the board
directly!