move: add Capture constructor
If I ever want to score capture moves differently, it's now trivial!
Just have to use the constructor.

This required a slight rework to `obstructed_distance()` to make it now
specify when a distance is to a capture - but with a custom type
Distance, this was trivial! The `legal_sliding_moves()` function is now
also much cleaner - it uses a helper function for the non-capturing
moves, and creates the capturing move itself if necessary.