move: rename -> `action`
Now, we have a Move, and an Action. A Move is dead simple - just from
one position to another position. An Action is also dead simple - just
wraps a Move with a specific record. Akin to the split between Square
and Piece, this lets a function only deal with the relevant part. For
example, the Move module has a `to_string` for simply displaying
something like "a1 -> a2" - which the Action module can wrap to add
something like "Capture a1 -> a2". This leads to more modular code, and
better separation of concerns.