piece: represent as a type and a color
Much better for pattern matching, since we can now match on the specific
type of piece, without always having to do `Pawn(_)` and other annoying
syntax. Did of course require a lot of refactoring in other files to use
this new standard. Notably, `promotable_pieces` is now
`promotable_piece_kinds`, and it returns a list of PieceKinds - which is
a kind of piece that isn't associated with a specific color. If the user
wants to associate color, they can do that themselves.