board: bring back `get_data` function
Just when I removed it, I found a use for it - iterating through
squares for a function I'm writing in the Game module. We don't need to
add the overhead of iterating through indices and getting the square for
each index - so keep this around. I still think `from_data` is silly,
though - so that's not coming back.