board/{index}?format: deduplicate logic
Now THIS is functional programming! The whole idea behind `board.format`
and `board.index_format` is that they abstract away the constant parts
of formatting a board (the line splitting, the labeling, etc. They let
you focus on the relationship between a single square and a string by
passing a function that handles this. Their only difference is whether
they care about the position on the board or not (so whether they use
`list.map` or `list.index_map`).

However, I was running into issues with duplicating logic between the
two. I try to not be stingy about DRY (Don't Repeat Yourself), because I
think doing it too much creates unnecessary abstractions. But, I got to
the point where I felt this would be useful to prevent the constant
copy/pasting. But, how could I do this?

Well, more functions! We add `create_formatter()`, which takes a board
and a function. Sounds familiar, right? The only difference is that this
is NOT a user function, but a mapper function. The user functions have a
signature like `fn(Square) -> String`, while the mapper functions have a
signature like `fn(List(Square)) -> List(String)`. This means
`board.format()` and `board.index_format` don't abstract away the actual
logic anymore - just the type of `map` function.