tests/generate: add `legal_moves` tests
Trying not to have super busy boards on here, because then the tests are
harder to check for accuracy. We make sure to check things like pieces
that would bump into each other, captures functioning and being properly
tracked, and no pieces of the correct color resulting in no legal moves.