legal/target: init
This is aimed to replace a Change, to save on memory. A Target does not
contain the starting position for a move, and is meant to be wrapped
with a type that holds only one origin value. This is aimed to prevent
so much duplicate data being used to store a move's starting position.