game: move `apply` logic into its own module
I know, I just did all that to make it possible to bring `apply` into
the game module. But I think that work was all still good. Part of my
problem with circular imports has been trying to make modules in the
object-oriented style - anything that works with a Game should be in the
Game module, etc. I'm trying to be more vibes-based, especially with the
new folders. A module is just where a collection of functions go when
they all serve one purpose. This makes circular imports way less common
and overall leads to small, modular files!