target: init
Taking the ideas from the `destination` branch, while not requiring a
complete rewrite. The core functionality of the `generate` module, for
getting all the moves from a given position, is now stored in the
`target` module - which, given some origin position, returns that origin
position ONCE, and a list of all the destinations from the origin. This
is a ton more efficient, since a list of moves requires storing the
origin for every single move.

However, the key difference from that branch is that we keep the
generate module around - we've put a ton of time into the infrastructure
around it, and we can't rewrite everything right now. However, now that
it's just a wrapper, we can access the target module directly for stuff
like check detection - and if I ever decide I want to put myself in
hell, I can rewrite everything.