targets: simplify API
Rather than making the target logic cater to the `generate` module, we
just change `generate.moves_from` to handle the origin logic directly!
Now the target module is more approachable, for direct access to the
targets without going through a tuple - which is probably what you
wanted anyways.

Oh, and we renamed the module from `target` -> `targets`, and renamed
the function so it's more grammatically intuitive. Targets from a
position? Yes please!