check: init
Simple check detection. We use the `target` logic for finding all the
legal moves of our enemy, and use that as our list of squares currently
attacked.

I thought originally that this would have a critical flaw, where it
wouldn't be able to detect certain kinds of check. However, I actually
don't think that'll be a problem! I thought that it might miss some
moves as being a move into check, because the enemy wouldn't consider
something a legal move until we had actually moved there. However, check
detection doesn't happen BEFORE making a move - it happens AFTER. We
apply each move and see whether it would put us into check - then filter
out the ones that would. This means that for any move, we'll try it, and
see whether that position is attacked. So, it should never do us wrong!

Also, I accidentally implemented *checkmate* detection for ages and only
realized I was overcomplicating it just now. That's a *much* harder
problem to solve, and thankfully I don't think we'll have to - at least
for now.