position: store as integer internally
Most of the data I'm working with is positions - the board is a
constant size, and it's not very nested - just an array of optional
Pieces, which have a Color. Moves, however, are much more nested - a
Move is a type of Change, and a Change maps one position to the other.
If we're going to be doing all sorts of operations with these, we should
represent them simply - as integers.

Some weird bugs - apparently, `board.get_pos` and `board.set_pos` have
been using the `player_index` functions the whole time! I must've never
noticed because of some other bug within the position module that
canceled it out. Regardless, it's fixed now - we don't even have a
`from_player_index` function anymore, it's useless. If you want an
index, you want it for operating on data.

The File and Rank modules are starting to feel quite useless - maybe
we'll bring them into the position module at some point. That's a change
for another day.