targets: add `to_pos`
Returns the positions that contain a friend who might be pinning our
current square, to prevent us from being in check! This is so we can
filter the pseudolegal moves for legal ones without needing to apply
every single move - only the ones that might actually leave us in check.