targets/to_pos: fix bug with friendly pieces
Before, we weren't flipping the color of the game - so when we generated
the sliding and knight targets, if they were a friendly piece, the
generation would decide that it couldn't kill a friendly piece, and
wouldn't generate that move. I don't know how this being missing didn't
break *more* things - but it's fixed now - and we properly flip the
color. Finally, we add a test using the game that I discovered this on.
Probably won't make a difference, but might as well.