generate: prevent more check mistakes
Running the bot against itself finds me a lot of these! This time, it
was for being within check. We were only applying moves that were
in the ring around the king - but another type of move that could take
us out of check is a move killing the assasin that's currently checking
us. In the process, we cleaned up the `targets` helper functions for
determining check - we now have `from_pos`, `enemies_attacking_pos`, and
`friends_attacking_pos`. This is a lot easier for me to conceptualize.