targets/passant: fix generation
Had some weird bugs with how we were handling rank and file index, since
I didn't realize they were 0-based. We add a test with the game that
made me realize the problem in the first place, and some helper
functions in the `position` module to explicitly get 1-based indexes.