algebraic: fix notation for en passant
We now properly show en passant characters to look like "exd5" (or its
equivalent on some other rank and file). To do this, we finally stripped
down the logic for pawn captures, instead relying on the Capture record
to see if something is actually a capture. Technically, it's one more
thing that can go wrong if something gets the wrong label - but I'll
hope it'll turn out okay.