generate: check for all line of fire squares
Previously, we considered "line of fire" squares as squares in the 1x1
ring around the king - but there can be a bishop or knight line that
also could be the line of fire. In fixing this, we cleaned up the
`generate` API - we now can directly get empty, friendly, and enemy
positions from the king's position, and they all use an internal helper
function. I wrote a better version, that used a custom type and didn't
require filtering at all - but I hit weird bugs with it, so let's stick
to what works.