generate: only apply if from attacked square
Big efficiency gain - when we're not in check (which is most of the
time), we now don't check every position that might be defending our
king - the position has to actually be attacked. We can do even better
by this, by only checking for the direct lines to the king - but this is
a big improvement already. We also should think about using similar
logic for when we're in check.