generate: filter pseudolegal moves in `moves_from`
Previously, we would only filter in the `legal_moves` function - which
meant that I was getting weird bugs with issues only cropping up when I
used one of the functions. Now, we move the filtering to `moves_from`,
so it's always handled! With this, we can bring in a test that was
previously commented out, since I didn't understand why it failed
inconsistently.