game/obstructed_distance: don't assume color
It's become useful to lie about the color of some piece, to see what we
could capture if there was an enemy there. This is why we have the
new `from_pos_as_piece` function. This propagated to almost all areas of
the `targets` module, since I do a great job at passing the piece
around. However, it wasn't passed for sliding pieces. Well, now it is -
and the `endangered_positions` function works completely as intended.