query/dangerous_destinations: init
Replacing the old `endangered_positions` function. This generates all
the squares that are dangerous to move to, since if we moved there, we
could be captured. To do this, I had to add a lot of ugliness to the
`targets` generation - ideally, we wouldn't have to do this, but we're
quite limited on time at this point, and I don't expect to have to touch
the module before this is due, so it'll do for now. No tests - forgive
me!